// src/api/generated/model/workoutExerciseResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Fitness App API
 * Comprehensive API for fitness application with support for workout plans, exercises, and progress tracking. API provides OAuth2 authentication via Keycloak or JWT tokens.
 */
import type { ExerciseDetailResponse } from './exerciseDetailResponse';
import type { ReferenceDataResponse } from './referenceDataResponse';
import type { WorkoutExerciseSetResponse } from './workoutExerciseSetResponse';

export interface WorkoutExerciseResponse {
  id?: number;
  workoutId?: number;
  order?: number;
  exercise?: ExerciseDetailResponse;
  workoutExerciseMetric?: ReferenceDataResponse;
  note?: string;
  workoutExerciseSets?: WorkoutExerciseSetResponse[];
}