// src/api/generated/model/workoutSessionInputRequestStatus.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Fitness App API
 * Comprehensive API for fitness application with support for workout plans, exercises, and progress tracking. API provides OAuth2 authentication via Keycloak or JWT tokens.
 */

export type WorkoutSessionInputRequestStatus = typeof WorkoutSessionInputRequestStatus[keyof typeof WorkoutSessionInputRequestStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const WorkoutSessionInputRequestStatus = {
  NOT_STARTED: 'NOT_STARTED',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
  SKIPPED: 'SKIPPED',
} as const;